+++ include/httpHandling.c
@@ -211,12 +211,12 @@ void handleHTTPConnection()
 
 	// initialize local variables
 	pthread_mutex_t *mutexes = setupMutexes(MAX_CONNECTION_COUNT * 2);
-	struct connectionResources connections[MAX_CONNECTION_COUNT];
+	struct connection_resources connections[MAX_CONNECTION_COUNT];
 
 	for(int i = 0; i < MAX_CONNECTION_COUNT; i++)
 	{
-		connections[i].mutex_data_from_clientSize = mutexes[i * 2];
-		connections[i].mutex_dataFromServerSize = mutexes[i * 2 + 1];
+		connections[i].mutex_data_from_client = mutexes[i * 2];
+		connections[i].mutex_data_from_server = mutexes[i * 2 + 1];
 	}
 
 	setupConnectionResources(connections, MAX_CONNECTION_COUNT, output_file_ptr);
@@ -225,42 +225,42 @@ void handleHTTPConnection()
 	// initialize variables for listening thread
 	pthread_t listeningThread;
 	int hostSocket = returnListeningSocket();
-	int acceptedSocket;
-	bool acceptedSocketPending;
-	bool shutdownListeningSocket;
-	pthread_mutex_t mutex_acceptedSocket = PTHREAD_MUTEX_INITIALIZER;
-	struct listeningThreadParameters listeningThreadArgs;
+	int accepted_socket;
+	bool accepted_socket_pending;
+	bool shutdownlistening_socket;
+	pthread_mutex_t mutex_accepted_socket = PTHREAD_MUTEX_INITIALIZER;
+	struct listening_thread_parameters listeningThreadArgs;
 
-	listeningThreadArgs.listeningSocket = hostSocket;
-	listeningThreadArgs.acceptedSocket = &acceptedSocket;
-	listeningThreadArgs.acceptedSocketPending = &acceptedSocketPending;
-	listeningThreadArgs.shutdown = &shutdownListeningSocket;
-	listeningThreadArgs.mutex_acceptedSocket = &mutex_acceptedSocket;
+	listeningThreadArgs.listening_socket = hostSocket;
+	listeningThreadArgs.accepted_socket = &accepted_socket;
+	listeningThreadArgs.accepted_socket_pending = &accepted_socket_pending;
+	listeningThreadArgs.shutdown = &shutdownlistening_socket;
+	listeningThreadArgs.mutex_accepted_socket = &mutex_accepted_socket;
 
 	// create listening thread
 	pthread_create(&listeningThread, NULL, listeningThreadFunction, &listeningThreadArgs);
 
-	while(!shutdownListeningSocket)
+	while(!shutdownlistening_socket)
 	{
 		if(connectionCount == MAX_CONNECTION_COUNT)
-			shutdownListeningSocket = true;
+			shutdownlistening_socket = true;
 
 		// there is a new connection pending
-		if(acceptedSocketPending && connectionCount < MAX_CONNECTION_COUNT)
+		if(accepted_socket_pending && connectionCount < MAX_CONNECTION_COUNT)
 		{
-			struct connectionResources *temp = &connections[connectionCount];
+			struct connection_resources *temp = &connections[connectionCount];
 			pthread_mutex_lock(&mutex_outputFile);
 			printf("[   main   ] new connection made\n");
 			fprintf(output_file_ptr, "[   main   ] new connection made\n");
 			pthread_mutex_unlock(&mutex_outputFile);
 
-			pthread_mutex_lock(&mutex_acceptedSocket);
-			temp->clientSocket = acceptedSocket;
-			acceptedSocketPending = false;
-			pthread_mutex_unlock(&mutex_acceptedSocket);
+			pthread_mutex_lock(&mutex_accepted_socket);
+			temp->client_socket = accepted_socket;
+			accepted_socket_pending = false;
+			pthread_mutex_unlock(&mutex_accepted_socket);
 			temp->data_from_client[0] = '\0';
-			temp->dataFromServer[0] = '\0';
-			int receiveLength = recv(temp->clientSocket, temp->data_from_client, BUFFER_SIZE, 0);
+			temp->data_from_server[0] = '\0';
+			int receiveLength = recv(temp->client_socket, temp->data_from_client, BUFFER_SIZE, 0);
 
 			if(receiveLength == -1)
 			{
@@ -272,7 +272,7 @@ void handleHTTPConnection()
 					printf("[   main   ] resetting connection...\n");
 					fprintf(output_file_ptr, "[   main   ] resetting connection...\n");
 					pthread_mutex_unlock(&mutex_outputFile);
-					close(temp->clientSocket);
+					close(temp->client_socket);
 					continue;
 				}
 
@@ -292,11 +292,11 @@ void handleHTTPConnection()
 			fprintf(output_file_ptr, "[   main   ] received %d bytes from client %d\n", receiveLength, connectionCount);
 			dump(temp->data_from_client, receiveLength, output_file_ptr);
 			pthread_mutex_unlock(&mutex_outputFile);
-			temp->data_from_clientSize = receiveLength;
+			temp->data_from_client_size = receiveLength;
 
 			// get information about server
-			char destinationName[DESTINATION_NAME_LENGTH + 1];
-			int functionResult = getDestinationName(temp->data_from_client, destinationName, output_file_ptr);
+			char destination_name[DESTINATION_NAME_LENGTH + 1];
+			int functionResult = getDestinationName(temp->data_from_client, destination_name, output_file_ptr);
 
 			if(functionResult == -1)
 			{
@@ -304,18 +304,18 @@ void handleHTTPConnection()
 				printf("[   main   ] error finding host string\n");
 				fprintf(output_file_ptr, "[   main   ] error finding host string\n");
 				pthread_mutex_unlock(&mutex_outputFile);
-				close(temp->clientSocket);
+				close(temp->client_socket);
 				continue;
 			}
 
 			char destinationPort[DESTINATION_PORT_LENGTH + 1] = LISTENING_PORT;
 
-			struct addrinfo destinationAddressInformation = returnDestinationAddressInfo(destinationName, destinationPort, output_file_ptr);
+			struct addrinfo destinationAddressInformation = returnDestinationAddressInfo(destination_name, destinationPort, output_file_ptr);
 
 			// not whitelisted
-			if(!isWhitelisted(whitelist, destinationName, destinationPort, destinationAddressInformation))
+			if(!isWhitelisted(whitelist, destination_name, destinationPort, destinationAddressInformation))
 			{
-				temp->serverSocket = 0;
+				temp->server_socket = 0;
 				threadFunction = &blacklistedThreadFunction;
 			}
 			// whitelisted
@@ -324,7 +324,7 @@ void handleHTTPConnection()
 				threadFunction = &whitelistedThreadFunction;
 
 				// create socket to destination
-				temp->serverSocket = returnSocketToServer(destinationAddressInformation);
+				temp->server_socket = returnSocketToServer(destinationAddressInformation);
 				pthread_mutex_lock(&mutex_outputFile);
 				printf("[   main   ] established TCP connection with server\n");
 				fprintf(output_file_ptr, "[   main   ] established TCP connection with server\n");
@@ -332,8 +332,8 @@ void handleHTTPConnection()
 			}
 
 			// create threads
-			pthread_create(&connections[connectionCount].clientThread, NULL, threadFunction, &connections[connectionCount].clientArgs);
-			pthread_create(&connections[connectionCount].serverThread, NULL, threadFunction, &connections[connectionCount].serverArgs);
+			pthread_create(&connections[connectionCount].clientThread, NULL, threadFunction, &connections[connectionCount].client_arguments);
+			pthread_create(&connections[connectionCount].serverThread, NULL, threadFunction, &connections[connectionCount].server_arguments);
 
 			connectionCount++;
 		}
@@ -352,7 +352,7 @@ void handleHTTPConnection()
 	pthread_mutex_destroy(&mutex_outputFile);
 }
 
-bool isWhitelisted(const struct whitelistStructure whitelist, const char* destinationName, const char* destinationPort, const struct addrinfo addressInfo)
+bool isWhitelisted(const struct whitelist_structure whitelist, const char* destination_name, const char* destinationPort, const struct addrinfo addressInfo)
 {
 	char destinationAddressString[INET_ADDRSTRLEN];
 	struct sockaddr_in destinationAddress_in = *(struct sockaddr_in *)addressInfo.ai_addr;
@@ -368,14 +368,14 @@ bool isWhitelisted(const struct whitelistStructure whitelist, const char* destin
 	}
 
 	// test for hostname match
-	for(int i = 0;i<whitelist.hostnameCount;i++)
+	for(int i = 0;i<whitelist.hostname_count;i++)
 	{
-		if(strcmp(destinationName, whitelist.hostnames[i]) == 0)
+		if(strcmp(destination_name, whitelist.hostnames[i]) == 0)
 			return true;
 	}
 
 	// test for port number match
-	for(int i = 0;i<whitelist.portCount;i++)
+	for(int i = 0;i<whitelist.port_count;i++)
 	{
 		if(strcmp(destinationPort, whitelist.ports[i]) == 0)
 			return true;
@@ -388,7 +388,7 @@ bool isWhitelisted(const struct whitelistStructure whitelist, const char* destin
 /* create, bind, and return a listening socket */
 int returnListeningSocket()
 {
-	char function_name[] = "returnListeningSocket";
+	char function_name[] = "returnlistening_socket";
 	struct addrinfo hostAddrHint, *hostResult;
 	int hostSocket;
 
@@ -405,7 +405,7 @@ int returnListeningSocket()
 
 	int yes = 1;
 
-	if(setsockopt(hostSocket, SOL_SOCKET, SO_REUSEADDR, &yes, sizeof(yes)))
+	if(setsockopt(hostSocket, SOL_SOCKET, SO_REUSEADDR, &yes, sizeof(yes)) == -1)
 		fatal("setsockopt", function_name, stdout);
 
 	if(bind(hostSocket, hostResult->ai_addr, hostResult->ai_addrlen) == -1)
@@ -421,7 +421,7 @@ int returnListeningSocket()
 /*
  * create, connect, and return a socket to the client
  */
-int returnSocketToClient(const int listeningSocket)
+int returnSocketToClient(const int listening_socket)
 {
 	struct sockaddr clientAddress;
 	socklen_t sin_size = sizeof(struct sockaddr);
@@ -429,7 +429,7 @@ int returnSocketToClient(const int listeningSocket)
 
 	while(1)
 	{
-		socketToClient = accept(listeningSocket, &clientAddress, &sin_size);
+		socketToClient = accept(listening_socket, &clientAddress, &sin_size);
 
 		if(socketToClient == -1)
 		{
@@ -453,46 +453,46 @@ int returnSocketToClient(const int listeningSocket)
 
 /* extract the destination name string from the HTTP request */
 /* returns offset of name from start of data, or on error:	-1 when error finding the host string */
-int getDestinationName(const unsigned char *receivedData, char *destinationNameBuffer, FILE *output_file_ptr)
+int getDestinationName(const unsigned char *receivedData, char *destination_nameBuffer, FILE *output_file_ptr)
 {
-	char *destinationNameStart, *destinationNameEnd;
-	int destinationNameLength;
+	char *destination_nameStart, *destination_nameEnd;
+	int destination_nameLength;
 	int domainNameIndex = 0;
 
-	destinationNameStart = strstr((char *)receivedData, "Host: ");
+	destination_nameStart = strstr((char *)receivedData, "Host: ");
 
-	if(destinationNameStart == NULL)
+	if(destination_nameStart == NULL)
 		return -1;
 
-	destinationNameStart += 6;
-	destinationNameEnd = NULL;
+	destination_nameStart += 6;
+	destination_nameEnd = NULL;
 
-	while(destinationNameEnd == NULL)
+	while(destination_nameEnd == NULL)
 	{
 		// reached end of file without finding domain
 		if(domainNameIndex == DOMAIN_NAME_COUNT)
 			return -1;
 
-		destinationNameEnd = strstr(destinationNameStart, domainNames[domainNameIndex]);
+		destination_nameEnd = strstr(destination_nameStart, domainNames[domainNameIndex]);
 		domainNameIndex++;
 	}
 
 	// TODO: change for domain names that aren't exactly 3 characters long
-	destinationNameEnd += 4;
-	destinationNameLength = destinationNameEnd - destinationNameStart;
+	destination_nameEnd += 4;
+	destination_nameLength = destination_nameEnd - destination_nameStart;
 
-	strncpy(destinationNameBuffer, destinationNameStart, destinationNameLength);
-	destinationNameBuffer[destinationNameLength] = '\0';
+	strncpy(destination_nameBuffer, destination_nameStart, destination_nameLength);
+	destination_nameBuffer[destination_nameLength] = '\0';
 
 	pthread_mutex_lock(&mutex_outputFile);
-	printf("destination name is: %s\n", destinationNameBuffer);
-	fprintf(output_file_ptr, "destination name is: %s\n", destinationNameBuffer);
+	printf("destination name is: %s\n", destination_nameBuffer);
+	fprintf(output_file_ptr, "destination name is: %s\n", destination_nameBuffer);
 	pthread_mutex_unlock(&mutex_outputFile);
-	return (char *)receivedData - destinationNameStart;
+	return (char *)receivedData - destination_nameStart;
 }
 
 /* get additional information about the destination */
-struct addrinfo returnDestinationAddressInfo(const char *destinationName, const char *destinationPort, FILE *output_file_ptr)
+struct addrinfo returnDestinationAddressInfo(const char *destination_name, const char *destinationPort, FILE *output_file_ptr)
 {
 	char function_name[] = "returnDestinationAddressInfo";
 	struct addrinfo destinationAddressHint, *destinationAddressResult;
@@ -500,7 +500,7 @@ struct addrinfo returnDestinationAddressInfo(const char *destinationName, const 
 	destinationAddressHint.ai_family = AF_INET;
 	destinationAddressHint.ai_socktype = SOCK_STREAM;
 
-	if(getaddrinfo(destinationName, destinationPort, &destinationAddressHint, &destinationAddressResult) != 0)
+	if(getaddrinfo(destination_name, destinationPort, &destinationAddressHint, &destinationAddressResult) != 0)
 		fatal("getting address information for the destination", function_name, stdout);
 
 	char destinationAddressString[INET_ADDRSTRLEN];
@@ -549,26 +549,26 @@ void *whitelistedThreadFunction(void *args)
 {
 #define CONNECTION_TIMEOUT_VALUE 2
 	// set up local variables with argument
-	struct threadParameters parameters = *(struct threadParameters *)args;
+	struct thread_parameters parameters = *(struct thread_parameters *)args;
 	// connection info
 	const int socket = *parameters.socket;
-	const int ID = parameters.connectionID;
-	char connectedTo[NAME_LENGTH];
-	memcpy(connectedTo, parameters.connectedTo, NAME_LENGTH);
+	const int ID = parameters.connection_ID;
+	char connected_to[NAME_LENGTH];
+	memcpy(connected_to, parameters.connected_to, NAME_LENGTH);
 	bool *shutdown = parameters.shutdown;
 	// read/write buffer info
-	int *readBufferSize = parameters.readBufferSize;
-	int *writeBufferSize = parameters.writeBufferSize;
-	unsigned char *writeBuffer = parameters.writeBuffer;
-	const unsigned char *readBuffer = parameters.readBuffer;
+	int *read_buffer_size = parameters.read_buffer_size;
+	int *write_buffer_size = parameters.write_buffer_size;
+	unsigned char *write_buffer = parameters.write_buffer;
+	const unsigned char *read_buffer = parameters.read_buffer;
 	// file pointers
 	FILE *output_file_ptr = parameters.output_file_ptr;
 	FILE *debug_file_ptr = parameters.debug_file_ptr;
 	// mutex locks
-	pthread_mutex_t *mutex_writeBuffer = parameters.mutex_writeBufferSize;
-	pthread_mutex_t *mutex_readBuffer = parameters.mutex_readBufferSize;
+	pthread_mutex_t *mutex_write_buffer = parameters.mutex_write_buffer;
+	pthread_mutex_t *mutex_read_buffer = parameters.mutex_read_buffer;
 
-	unsigned char tempReadBuffer[BUFFER_SIZE + 1];
+	unsigned char tempread_buffer[BUFFER_SIZE + 1];
 	ssize_t recvResult;
 
 	// set up timeout
@@ -585,13 +585,13 @@ void *whitelistedThreadFunction(void *args)
 		{
 			*shutdown = true;
 			pthread_mutex_lock(&mutex_outputFile);
-			printf("[%d - %s] Terminating: idle connection\n", ID, connectedTo);
-			fprintf(debug_file_ptr, "[%d - %s] Terminating: idle connection\n", ID, connectedTo);
+			printf("[%d - %s] Terminating: idle connection\n", ID, connected_to);
+			fprintf(debug_file_ptr, "[%d - %s] Terminating: idle connection\n", ID, connected_to);
 			pthread_mutex_unlock(&mutex_outputFile);
 			pthread_exit(NULL);
 		}
 
-		recvResult = recv(socket, tempReadBuffer, BUFFER_SIZE, 0);
+		recvResult = recv(socket, tempread_buffer, BUFFER_SIZE, 0);
 
 		// error reading data
 		if(recvResult == -1)
@@ -600,8 +600,8 @@ void *whitelistedThreadFunction(void *args)
 			{
 				*shutdown = true;
 				pthread_mutex_lock(&mutex_outputFile);
-				printf("[%d - %s] Terminating: Error reading data.\nErrno: %d\n", ID, connectedTo, errno);
-				fprintf(debug_file_ptr, "[%d - %s] Terminating: Error reading data.\nErrno: %d\n", ID, connectedTo, errno);
+				printf("[%d - %s] Terminating: Error reading data.\nErrno: %d\n", ID, connected_to, errno);
+				fprintf(debug_file_ptr, "[%d - %s] Terminating: Error reading data.\nErrno: %d\n", ID, connected_to, errno);
 				pthread_mutex_unlock(&mutex_outputFile);
 				pthread_exit(NULL);
 			}
@@ -620,59 +620,59 @@ void *whitelistedThreadFunction(void *args)
 			{
 				*shutdown = true;
 				pthread_mutex_lock(&mutex_outputFile);
-				printf("[%d - %s] Terminating: 0 bytes received\n", ID, connectedTo);
-				fprintf(debug_file_ptr, "[%d - %s] Terminating: 0 bytes received\n", ID, connectedTo);
+				printf("[%d - %s] Terminating: 0 bytes received\n", ID, connected_to);
+				fprintf(debug_file_ptr, "[%d - %s] Terminating: 0 bytes received\n", ID, connected_to);
 				pthread_mutex_unlock(&mutex_outputFile);
 				pthread_exit(NULL);
 			}
 
 			pthread_mutex_lock(&mutex_outputFile);
-			printf("[%d - %s] Read %zd bytes\n", ID, connectedTo, recvResult);
-			fprintf(debug_file_ptr, "[%d - %s] Read %zd bytes\n", ID, connectedTo, recvResult);
-			fprintf(output_file_ptr, "[%d - %s] Read %zd bytes\n", ID, connectedTo, recvResult);
-			dump(tempReadBuffer, recvResult, output_file_ptr);
+			printf("[%d - %s] Read %zd bytes\n", ID, connected_to, recvResult);
+			fprintf(debug_file_ptr, "[%d - %s] Read %zd bytes\n", ID, connected_to, recvResult);
+			fprintf(output_file_ptr, "[%d - %s] Read %zd bytes\n", ID, connected_to, recvResult);
+			dump(tempread_buffer, recvResult, output_file_ptr);
 			pthread_mutex_unlock(&mutex_outputFile);
 
 			// wait until buffer is empty before writing to it
-			while(*writeBufferSize != 0) {};
+			while(*write_buffer_size != 0) {};
 
 			// write to buffer and change buffer size
-			pthread_mutex_lock(mutex_writeBuffer);
-			memcpy(writeBuffer, tempReadBuffer, recvResult);
-			*writeBufferSize = recvResult;
-			pthread_mutex_unlock(mutex_writeBuffer);
+			pthread_mutex_lock(mutex_write_buffer);
+			memcpy(write_buffer, tempread_buffer, recvResult);
+			*write_buffer_size = recvResult;
+			pthread_mutex_unlock(mutex_write_buffer);
 
 			pthread_mutex_lock(&mutex_outputFile);
-			printf("[%d - %s] Wrote %zd bytes\n", ID, connectedTo, recvResult);
-			fprintf(debug_file_ptr, "[%d - %s] Wrote %zd bytes\n", ID, connectedTo, recvResult);
+			printf("[%d - %s] Wrote %zd bytes\n", ID, connected_to, recvResult);
+			fprintf(debug_file_ptr, "[%d - %s] Wrote %zd bytes\n", ID, connected_to, recvResult);
 			pthread_mutex_unlock(&mutex_outputFile);
 		}
 
 		// if there is data in the read buffer, send it
-		if(*readBufferSize != 0)
+		if(*read_buffer_size != 0)
 		{
-			if(sendString(socket, readBuffer, *readBufferSize) == 0)
+			if(sendString(socket, read_buffer, *read_buffer_size) == 0)
 			{
 				pthread_mutex_lock(&mutex_outputFile);
 				*shutdown = true;
-				printf("[%d - %s] Terminating: error sending data\n", ID, connectedTo);
-				fprintf(debug_file_ptr, "[%d - %s] Terminating: error sending data\n", ID, connectedTo);
+				printf("[%d - %s] Terminating: error sending data\n", ID, connected_to);
+				fprintf(debug_file_ptr, "[%d - %s] Terminating: error sending data\n", ID, connected_to);
 				pthread_mutex_unlock(&mutex_outputFile);
 				pthread_exit(NULL);
 			}
 
 			pthread_mutex_lock(&mutex_outputFile);
-			printf("[%d - %s] Sent data\n", ID, connectedTo);
-			fprintf(debug_file_ptr, "[%d - %s] Sent data\n", ID, connectedTo);
+			printf("[%d - %s] Sent data\n", ID, connected_to);
+			fprintf(debug_file_ptr, "[%d - %s] Sent data\n", ID, connected_to);
 			pthread_mutex_unlock(&mutex_outputFile);
 
-			pthread_mutex_lock(mutex_readBuffer);
-			*readBufferSize = 0;
-			pthread_mutex_unlock(mutex_readBuffer);
+			pthread_mutex_lock(mutex_read_buffer);
+			*read_buffer_size = 0;
+			pthread_mutex_unlock(mutex_read_buffer);
 
 			pthread_mutex_lock(&mutex_outputFile);
-			printf("[%d - %s] Set buffer to empty\n", ID, connectedTo);
-			fprintf(debug_file_ptr, "[%d - %s] Set buffer to empty\n", ID, connectedTo);
+			printf("[%d - %s] Set buffer to empty\n", ID, connected_to);
+			fprintf(debug_file_ptr, "[%d - %s] Set buffer to empty\n", ID, connected_to);
 			pthread_mutex_unlock(&mutex_outputFile);
 		}
 	}
@@ -680,64 +680,64 @@ void *whitelistedThreadFunction(void *args)
 	// clean up code
 	close(socket);
 	pthread_mutex_lock(&mutex_outputFile);
-	printf("[%d - %s] Terminating: shutdown variable set\n", ID, connectedTo);
-	fprintf(debug_file_ptr, "[%d - %s] Terminating: shutdown variable set\n", ID, connectedTo);
+	printf("[%d - %s] Terminating: shutdown variable set\n", ID, connected_to);
+	fprintf(debug_file_ptr, "[%d - %s] Terminating: shutdown variable set\n", ID, connected_to);
 	pthread_mutex_unlock(&mutex_outputFile);
 	pthread_exit(NULL);
 }
 
 void *listeningThreadFunction(void *args)
 {
-	struct listeningThreadParameters parameter = *(struct listeningThreadParameters *)args;
-	int listeningSocket = parameter.listeningSocket;
-	int *acceptedSocket = parameter.acceptedSocket;
-	bool *acceptedSocketPending = parameter.acceptedSocketPending;
+	struct listening_thread_parameters parameter = *(struct listening_thread_parameters *)args;
+	int listening_socket = parameter.listening_socket;
+	int *accepted_socket = parameter.accepted_socket;
+	bool *accepted_socket_pending = parameter.accepted_socket_pending;
 	bool *shutdown = parameter.shutdown;
-	pthread_mutex_t *mutex_acceptedSocket = parameter.mutex_acceptedSocket;
+	pthread_mutex_t *mutex_accepted_socket = parameter.mutex_accepted_socket;
 
-	int tempAcceptedSocket = 0;
+	int tempaccepted_socket = 0;
 
 	printf("[ listener ] Listening on port %s\n", LISTENING_PORT);
 
 	while(!(*shutdown))
 	{
-		if(tempAcceptedSocket == 0)
-			tempAcceptedSocket = returnSocketToClient(listeningSocket);
+		if(tempaccepted_socket == 0)
+			tempaccepted_socket = returnSocketToClient(listening_socket);
 
-		if(tempAcceptedSocket == -2)
+		if(tempaccepted_socket == -2)
 		{
 			printf("[ listener ] error while accepting connection\n");
 			*shutdown = true;
 			continue;
 		}
 
-		pthread_mutex_lock(mutex_acceptedSocket);
+		pthread_mutex_lock(mutex_accepted_socket);
 
-		if(!(*acceptedSocketPending))
+		if(!(*accepted_socket_pending))
 		{
-			*acceptedSocket = tempAcceptedSocket;
-			*acceptedSocketPending = true;
-			tempAcceptedSocket = 0;
+			*accepted_socket = tempaccepted_socket;
+			*accepted_socket_pending = true;
+			tempaccepted_socket = 0;
 		}
 
-		pthread_mutex_unlock(mutex_acceptedSocket);
+		pthread_mutex_unlock(mutex_accepted_socket);
 
 	}
 
-	if(tempAcceptedSocket != 0)
-		close(tempAcceptedSocket);
+	if(tempaccepted_socket != 0)
+		close(tempaccepted_socket);
 
-	pthread_mutex_lock(mutex_acceptedSocket);
+	pthread_mutex_lock(mutex_accepted_socket);
 
-	if(!(*acceptedSocketPending))
-		close(*acceptedSocket);
+	if(!(*accepted_socket_pending))
+		close(*accepted_socket);
 
-	pthread_mutex_unlock(mutex_acceptedSocket);
-	close(listeningSocket);
+	pthread_mutex_unlock(mutex_accepted_socket);
+	close(listening_socket);
 	pthread_exit(NULL);
 }
 
-void cleanupConnections(struct connectionResources *conRes, int connectionCount)
+void cleanupConnections(struct connection_resources *conRes, int connectionCount)
 {
 	void *retval;
 	int result;
@@ -782,28 +782,28 @@ void getRequestedObject(const unsigned char *requestMessage, char *requestedObje
 	}
 }
 
-int sendResponse(int socket, const int options, const char* fileType, char* writeBuffer, const struct HTTPResponse* response, FILE* output_file_ptr)
+int sendResponse(int socket, const int options, const char* fileType, char* write_buffer, const struct HTTP_response* response, FILE* output_file_ptr)
 {
 #define FILE_READ_BUFFER_SIZE 100
-	memset(writeBuffer, 0, 	BUFFER_SIZE);
-	strcat(writeBuffer, response->responseVersion);
-	strcat(writeBuffer, " ");
-	strcat(writeBuffer, response->statusCode);
-	strcat(writeBuffer, "\r\n");
+	memset(write_buffer, 0, 	BUFFER_SIZE);
+	strcat(write_buffer, response->response_version);
+	strcat(write_buffer, " ");
+	strcat(write_buffer, response->status_code);
+	strcat(write_buffer, "\r\n");
 	for(int i = 0;i<RESPONSE_HEADER_COUNT;i++)
 	{
-		if(response->headers[i].headerName == NULL)
+		if(response->headers[i].header_name == NULL)
 			break;
 		else
 		{
-			strcat(writeBuffer, response->headers[i].headerName);
-			strcat(writeBuffer, ": ");
-			strcat(writeBuffer, response->headers[i].headerData);
-			strcat(writeBuffer, "\r\n");
+			strcat(write_buffer, response->headers[i].header_name);
+			strcat(write_buffer, ": ");
+			strcat(write_buffer, response->headers[i].header_data);
+			strcat(write_buffer, "\r\n");
 		}
 	}
-	strcat(writeBuffer, "\r\n");
-	int writeBufferSize = strlen(writeBuffer);
+	strcat(write_buffer, "\r\n");
+	int write_buffer_size = strlen(write_buffer);
 	FILE* inputFile;
 	if((options & RESPONSE_NO_PAYLOAD) == 0)
 	{
@@ -817,18 +817,18 @@ int sendResponse(int socket, const int options, const char* fileType, char* writ
 		else
 			return -1;
 
-		char fileReadBuffer[FILE_READ_BUFFER_SIZE];
+		char fileread_buffer[FILE_READ_BUFFER_SIZE];
 		size_t bytesRead;
-		while((bytesRead = fread(fileReadBuffer, 1, FILE_READ_BUFFER_SIZE, inputFile)) > 0)
+		while((bytesRead = fread(fileread_buffer, 1, FILE_READ_BUFFER_SIZE, inputFile)) > 0)
 		{
-			memcpy(writeBuffer + writeBufferSize, fileReadBuffer, bytesRead);
-			writeBufferSize += bytesRead;
+			memcpy(write_buffer + write_buffer_size, fileread_buffer, bytesRead);
+			write_buffer_size += bytesRead;
 		}
 	}
 
-	fprintf(output_file_ptr, "Send response with size %d:\n", writeBufferSize);
-	dump((unsigned char*)writeBuffer, writeBufferSize, output_file_ptr);
-	write(socket, writeBuffer, writeBufferSize);
+	fprintf(output_file_ptr, "Send response with size %d:\n", write_buffer_size);
+	dump((unsigned char*)write_buffer, write_buffer_size, output_file_ptr);
+	write(socket, write_buffer, write_buffer_size);
 	fclose(inputFile);
 	return 0;
 }
@@ -838,16 +838,16 @@ void* blacklistedThreadFunction(void* args)
 #define REQUESTED_OBJECT_NAME_LENGTH 30
 #define REQUESTED_OBJECT_TYPE_LENGTH 7
 	// set up local variables with argument
-	struct threadParameters parameters = *(struct threadParameters *)args;
+	struct thread_parameters parameters = *(struct thread_parameters *)args;
 	const int socket = *parameters.socket;
 	// terminate if server thread
 	if(socket == 0)
 		pthread_exit(NULL);
-	const int ID = parameters.connectionID;
+	const int ID = parameters.connection_ID;
 	bool *shutdown = parameters.shutdown;
 	// read/write buffer info
-	unsigned char *data_from_client = parameters.writeBuffer;
-	unsigned char *dataToClient = parameters.readBuffer;
+	unsigned char *data_from_client = parameters.write_buffer;
+	unsigned char *dataToClient = parameters.read_buffer;
 	// file pointers
 	FILE *output_file_ptr = parameters.output_file_ptr;
 	FILE *debug_file_ptr = parameters.debug_file_ptr;
@@ -857,7 +857,7 @@ void* blacklistedThreadFunction(void* args)
 	int requestType = 0;
 	char requestedObject[REQUESTED_OBJECT_NAME_LENGTH];
 	char requestedObjectType[REQUESTED_OBJECT_TYPE_LENGTH];
-	static struct HTTPResponse* defaultResponse = NULL;
+	static struct HTTP_response* defaultResponse = NULL;
 	if(defaultResponse == NULL)
 		setupResponse(&defaultResponse, 0);
 	int packetCount = 0;
@@ -965,7 +965,7 @@ int getHTTPRequestType(const char* receivedData)
 		return 0;
 }
 
-void setupResponse(struct HTTPResponse** destination, int options)
+void setupResponse(struct HTTP_response** destination, int options)
 {
 #define SERVER_HEADER_DEFAULT "nginx/1.18.0 (Ubuntu)\0"
 #define DATE_HEADER_DEFAULT "Wed, 29 Jan 2025 23:45:35 GMT\0"
@@ -977,56 +977,56 @@ void setupResponse(struct HTTPResponse** destination, int options)
 #define XCONTENTTYPEOPTIONS_HEADER_DEFAULT "nosniff\0"
 #define CONTENTENCODING_HEADER_DEFAULT "gzip\0"
 
-	struct HTTPResponse* response  = (struct HTTPResponse*)malloc(sizeof(struct HTTPResponse) + (RESPONSE_HEADER_COUNT*sizeof(struct header)));
+	struct HTTP_response* response  = (struct HTTP_response*)malloc(sizeof(struct HTTP_response) + (RESPONSE_HEADER_COUNT*sizeof(struct header)));
 	*destination = response;
-	strcpy(response->responseVersion, "HTTP/1.1\0");
-	strcpy(response->statusCode, "200 OK\0");
+	strcpy(response->response_version, "HTTP/1.1\0");
+	strcpy(response->status_code, "200 OK\0");
 
-	response->headers[0].headerName = (char*)malloc(sizeof("Server\0"));
-	strcpy(response->headers[0].headerName, "Server\0");
-	response->headers[0].headerData = (char*)malloc(sizeof(SERVER_HEADER_DEFAULT));
-	strcpy(response->headers[0].headerData, SERVER_HEADER_DEFAULT);
+	response->headers[0].header_name = (char*)malloc(sizeof("Server\0"));
+	strcpy(response->headers[0].header_name, "Server\0");
+	response->headers[0].header_data = (char*)malloc(sizeof(SERVER_HEADER_DEFAULT));
+	strcpy(response->headers[0].header_data, SERVER_HEADER_DEFAULT);
 
-	response->headers[1].headerName = (char*)malloc(sizeof("Date\0"));
-	strcpy(response->headers[1].headerName, "Date\0");
-	response->headers[1].headerData = (char*)malloc(sizeof(DATE_HEADER_DEFAULT));
-	strcpy(response->headers[1].headerData, DATE_HEADER_DEFAULT);
+	response->headers[1].header_name = (char*)malloc(sizeof("Date\0"));
+	strcpy(response->headers[1].header_name, "Date\0");
+	response->headers[1].header_data = (char*)malloc(sizeof(DATE_HEADER_DEFAULT));
+	strcpy(response->headers[1].header_data, DATE_HEADER_DEFAULT);
 
-	response->headers[2].headerName = (char*)malloc(sizeof("Content-Type\0"));
-	strcpy(response->headers[2].headerName, "Content-Type\0");
-	response->headers[2].headerData = (char*)malloc(sizeof(CONTENTTYPE_HEADER_DEFAULT));
-	strcpy(response->headers[2].headerData, CONTENTTYPE_HEADER_DEFAULT);
+	response->headers[2].header_name = (char*)malloc(sizeof("Content-Type\0"));
+	strcpy(response->headers[2].header_name, "Content-Type\0");
+	response->headers[2].header_data = (char*)malloc(sizeof(CONTENTTYPE_HEADER_DEFAULT));
+	strcpy(response->headers[2].header_data, CONTENTTYPE_HEADER_DEFAULT);
 
-	response->headers[3].headerName = (char*)malloc(sizeof("Last-Modified\0"));
-	strcpy(response->headers[3].headerName, "Last-Modified\0");
-	response->headers[3].headerData = (char*)malloc(sizeof(LASTMODIFIED_HEADER_DEFAULT));
-	strcpy(response->headers[3].headerData, LASTMODIFIED_HEADER_DEFAULT);
+	response->headers[3].header_name = (char*)malloc(sizeof("Last-Modified\0"));
+	strcpy(response->headers[3].header_name, "Last-Modified\0");
+	response->headers[3].header_data = (char*)malloc(sizeof(LASTMODIFIED_HEADER_DEFAULT));
+	strcpy(response->headers[3].header_data, LASTMODIFIED_HEADER_DEFAULT);
 
-	response->headers[4].headerName = (char*)malloc(sizeof("Connection\0"));
-	strcpy(response->headers[4].headerName, "Connection\0");
-	response->headers[4].headerData = (char*)malloc(sizeof(CONNECTION_HEADER_DEFAULT));
-	strcpy(response->headers[4].headerData, CONNECTION_HEADER_DEFAULT);
+	response->headers[4].header_name = (char*)malloc(sizeof("Connection\0"));
+	strcpy(response->headers[4].header_name, "Connection\0");
+	response->headers[4].header_data = (char*)malloc(sizeof(CONNECTION_HEADER_DEFAULT));
+	strcpy(response->headers[4].header_data, CONNECTION_HEADER_DEFAULT);
 
-	response->headers[5].headerName = (char*)malloc(sizeof("ETag\0"));
-	strcpy(response->headers[5].headerName, "ETag\0");
-	response->headers[5].headerData = (char*)malloc(sizeof(ETAG_HEADER_DEFAULT));
-	strcpy(response->headers[5].headerData, ETAG_HEADER_DEFAULT);
+	response->headers[5].header_name = (char*)malloc(sizeof("ETag\0"));
+	strcpy(response->headers[5].header_name, "ETag\0");
+	response->headers[5].header_data = (char*)malloc(sizeof(ETAG_HEADER_DEFAULT));
+	strcpy(response->headers[5].header_data, ETAG_HEADER_DEFAULT);
 
-	response->headers[6].headerName = (char*)malloc(sizeof("Referrer-Policy\0"));
-	strcpy(response->headers[6].headerName, "Referrer-Policy\0");
-	response->headers[6].headerData = (char*)malloc(sizeof(REFERRERPOLICY_HEADER_DEFAULT));
-	strcpy(response->headers[6].headerData, REFERRERPOLICY_HEADER_DEFAULT);
+	response->headers[6].header_name = (char*)malloc(sizeof("Referrer-Policy\0"));
+	strcpy(response->headers[6].header_name, "Referrer-Policy\0");
+	response->headers[6].header_data = (char*)malloc(sizeof(REFERRERPOLICY_HEADER_DEFAULT));
+	strcpy(response->headers[6].header_data, REFERRERPOLICY_HEADER_DEFAULT);
 
-	response->headers[7].headerName = (char*)malloc(sizeof("X-Content-Type-Options\0"));
-	strcpy(response->headers[7].headerName, "X-Content-Type-Options\0");
-	response->headers[7].headerData = (char*)malloc(sizeof(XCONTENTTYPEOPTIONS_HEADER_DEFAULT));
-	strcpy(response->headers[7].headerData, XCONTENTTYPEOPTIONS_HEADER_DEFAULT);
+	response->headers[7].header_name = (char*)malloc(sizeof("X-Content-Type-Options\0"));
+	strcpy(response->headers[7].header_name, "X-Content-Type-Options\0");
+	response->headers[7].header_data = (char*)malloc(sizeof(XCONTENTTYPEOPTIONS_HEADER_DEFAULT));
+	strcpy(response->headers[7].header_data, XCONTENTTYPEOPTIONS_HEADER_DEFAULT);
 
-	response->headers[8].headerName = (char*)malloc(sizeof("Content-Encoding\0"));
-	strcpy(response->headers[8].headerName, "Content-Encoding\0");
-	response->headers[8].headerData = (char*)malloc(sizeof(CONTENTENCODING_HEADER_DEFAULT));
-	strcpy(response->headers[8].headerData, CONTENTENCODING_HEADER_DEFAULT);
+	response->headers[8].header_name = (char*)malloc(sizeof("Content-Encoding\0"));
+	strcpy(response->headers[8].header_name, "Content-Encoding\0");
+	response->headers[8].header_data = (char*)malloc(sizeof(CONTENTENCODING_HEADER_DEFAULT));
+	strcpy(response->headers[8].header_data, CONTENTENCODING_HEADER_DEFAULT);
 
-	response->headers[9].headerName = NULL;
-	response->headers[9].headerData = NULL;
+	response->headers[9].header_name = NULL;
+	response->headers[9].header_data = NULL;
 }